refactor(browser-extension): added type file
